added /services to fetch mock data and added fetch to onSubmit in flightForm
